refactoring stepper
