+++ pages/projects/[projectId]/carousel-with-stepper/index.tsx
@@ -1,20 +1,11 @@
 import { useState, useRef } from "react";
-import { MobileStepper, Box, Button } from "@mui/material";
-import { KeyboardArrowLeft, KeyboardArrowRight } from "@mui/icons-material";
-import Slider, { Settings } from "react-slick";
+import { Box } from "@mui/material";
+import Slider from "react-slick";
 import "slick-carousel/slick/slick.css";
 import "slick-carousel/slick/slick-theme.css";
-import {
-    MobileStepperButtonStyles,
-    MobileStepperStyles,
-    carouselWithStepperMainBoxStyles,
-    sliderStyles,
-} from "./styles";
-import {
-    PROJECT_DETAILS_SLIDER_BUTTON_LABEL_BACK,
-    PROJECT_DETAILS_SLIDER_BUTTON_LABEL_NEXT,
-} from "../../../../public/constants/text";
+import { carouselWithStepperMainBoxStyles } from "./styles";
 import ImagesSlider from "./slider";
+import Stepper from "./mobil-stepper";
 
 interface CarouselWithStepperProps {
     imagesPath: string[];
@@ -43,33 +34,11 @@ const CarouselWithStepper = ({ imagesPath }: CarouselWithStepperProps) => {
     return (
         <Box sx={carouselWithStepperMainBoxStyles}>
             <ImagesSlider imagesPath={imagesPath} sliderRef={sliderRef} />
-            <MobileStepper
-                steps={maxSteps}
-                position="static"
+            <Stepper
+                maxSteps={maxSteps}
                 activeStep={activeStep}
-                nextButton={
-                    <Button
-                        size="small"
-                        onClick={handleNext}
-                        disabled={activeStep === maxSteps - 1}
-                        sx={MobileStepperButtonStyles}
-                    >
-                        {PROJECT_DETAILS_SLIDER_BUTTON_LABEL_NEXT}
-                        <KeyboardArrowRight />
-                    </Button>
-                }
-                backButton={
-                    <Button
-                        size="small"
-                        onClick={handleBack}
-                        disabled={activeStep === 0}
-                        sx={MobileStepperButtonStyles}
-                    >
-                        <KeyboardArrowLeft />
-                        {PROJECT_DETAILS_SLIDER_BUTTON_LABEL_BACK}
-                    </Button>
-                }
-                sx={MobileStepperStyles}
+                handleNext={handleNext}
+                handleBack={handleBack}
             />
         </Box>
     );
