+++ pages/projects/[projectId]/carousel-with-stepper/mobil-stepper/styles.tsx
@@ -1,0 +1,13 @@
+export const MobileStepperButtonStyles = {
+    color: "#957b5f",
+};
+
+export const MobileStepperStyles = {
+    backgroundColor: "#2b2b2b",
+    "& .MuiMobileStepper-dotActive": {
+        backgroundColor: "#957b5f",
+    },
+    "& .MuiMobileStepper-dot": {
+        backgroundColor: "#d5d5d5",
+    },
+};
