+++ pages/projects/[projectId]/carousel-with-stepper/styles.tsx
@@ -2,25 +2,3 @@ export const carouselWithStepperMainBoxStyles = {
     width: "900px",
     backgroundColor: "#2b2b2b",
 };
-
-export const sliderStyles = {
-    height: "500px",
-    width: "700px",
-    display: "block",
-    maxWidth: "100%",
-    overflow: "hidden",
-};
-
-export const MobileStepperButtonStyles = {
-    color: "#957b5f",
-};
-
-export const MobileStepperStyles = {
-    backgroundColor: "#2b2b2b",
-    "& .MuiMobileStepper-dotActive": {
-        backgroundColor: "#957b5f",
-    },
-    "& .MuiMobileStepper-dot": {
-        backgroundColor: "#d5d5d5",
-    },
-};
